Refactored consts to separate files
