--- src/consts/consts.js
@@ -1,9 +1,0 @@
-export const GameType = {
-  ARTIST: `artist`,
-  GENRE: `genre`,
-};
-
-export const GameStep = {
-  START: -1,
-  FIRST: 0
-};
