+++ src/consts/game-step.js
@@ -1,0 +1,4 @@
+export default {
+  START: -1,
+  FIRST: 0
+};
