+++ src/consts/game-type.js
@@ -1,0 +1,4 @@
+export default {
+  ARTIST: `artist`,
+  GENRE: `genre`,
+};
