+++ src/consts/index.js
@@ -1,0 +1,2 @@
+export {default as GameType} from "./game-type";
+export {default as GameStep} from "./game-step";
